NO AUTO Strips refactor and periodic from full
