+++ core/src/test/java/apoc/periodic/BatchModeTest.java
