+++ core/src/test/java/apoc/refactor/util/PropertiesManagerTest.java
