+++ full/src/test/java/apoc/ApocSplitTest.java
@@ -17,14 +17,6 @@ import static org.junit.Assert.assertTrue;
  */
 public class ApocSplitTest {
     public static final List<String> PROCEDURES_FROM_CORE = List.of(
-        "apoc.periodic.truncate",
-        "apoc.periodic.list",
-        "apoc.periodic.commit",
-        "apoc.periodic.cancel",
-        "apoc.periodic.submit",
-        "apoc.periodic.repeat",
-        "apoc.periodic.countdown",
-        "apoc.periodic.iterate",
         "apoc.trigger.add",
         "apoc.trigger.remove",
         "apoc.trigger.removeAll",
@@ -171,24 +163,6 @@ public class ApocSplitTest {
         "apoc.warmup.run",
         "apoc.stats.degrees",
         "apoc.help",
-        "apoc.refactor.rename.label",
-        "apoc.refactor.rename.type",
-        "apoc.refactor.rename.nodeProperty",
-        "apoc.refactor.rename.typeProperty",
-        "apoc.refactor.extractNode",
-        "apoc.refactor.collapseNode",
-        "apoc.refactor.cloneNodes",
-        "apoc.refactor.cloneSubgraphFromPaths",
-        "apoc.refactor.cloneSubgraph",
-        "apoc.refactor.mergeNodes",
-        "apoc.refactor.mergeRelationships",
-        "apoc.refactor.setType",
-        "apoc.refactor.to",
-        "apoc.refactor.invert",
-        "apoc.refactor.from",
-        "apoc.refactor.normalizeAsBoolean",
-        "apoc.refactor.categorize",
-        "apoc.refactor.deleteAndReconnect",
         "apoc.convert.setJsonProperty",
         "apoc.convert.toTree",
         "apoc.neighbors.tohop",
